+++ src/app/app.module.ts
@@ -7,7 +7,7 @@ import { AppComponent } from "./app.component";
 
 import { MessageComponent, MessageListComponent, MessageModule } from "./messages";
 import { MessageFormComponent } from "./message-form";
-import { MessageService } from "../shared/services/index";
+import { MessageService } from "../shared/services";
 import { MessagePagerModule } from './messages/message-pager';
 import { MessagePagerComponent } from './messages/message-pager/message-pager.component';
 import { PagerItemComponent } from './messages/message-pager/pager-item-component';
